requirements.txt
opencv-python
numpy
matplotlib
torch
torchvision
torchaudio
ultralytics
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
pillow